fix logic in cover upload api

+++ b2d-venture/app/api/file/upload/[id]/cover/route.ts
@@ -20,11 +20,13 @@ export async function POST(req: Request, { params }) {
         if (!file) {
             return NextResponse.json({ error: 'No file uploaded.' }, { status: 400 });
         }
-        const business = Business.findById(id)
-        const oldUrl = business.coverimg
-        const oldUrlArray = oldUrl.split("/");
-        const oldName = oldUrlArray[oldUrlArray.lenght() - 1]
-        const result = await asset.deleteFile(`business/${id}/${oldName}`);
+        const business = await Business.findById(id)
+        const oldUrl =  business.coverimg || null
+        if (oldUrl){
+            const oldUrlArray = oldUrl.split("/");
+            const oldName = oldUrlArray[oldUrlArray.length - 1]
+            const result = await asset.deleteFile(`business/${id}/${oldName}`);
+        }
         const filePath = `business/${id}/${file.name}`
         const url = await asset.getPublicUrl(filePath)
         business.coverimg = url
